add Upstart 1.8 job file to smartkey-hun
:Release Notes:
Fix is copy/modify of existing upstart script ("smartkey.start" in src)
to meet requirements for Upstart 1.8

:Detailed Notes:
As the existing script did not use anything that would fail on Upstart
1.8, the fix is really re-arranging and cleaning that script to meet
our requirements for Upstart 1.8 job files.

:Testing Performed:
miniBAT will be run on a build ~240 based local build with Upstart 1.8
included.  For this component, my testing will be limited to making
sure that it is properly started and that it restarts if manually
killed.

pre-testing for this fix is to locally build the component and inspect
the output directory:
starfish/BUILD-goldfinger/work/goldfinger-starfish-linux-gnueabi/ \
    smartkey-hun/3.0.0-6-r4/image/etc/init and .../etc/event.d
to verify that the correct scripts are installed in the right spots.
The ipk file is also checked to make sure that the files are going to
be installed properly:
    deploy-ipks/goldfinger/smartkey-hun_3.0.0-6-r4_goldfinger.ipk

:QA Notes:

:Issues Addressed:
[GF-10827] add Upstart 1.8 script to smartkey-hun
  is required by:
[GF-10829] smartkey-hun: modify bb file to install new Upstart job file

Open-webOS-DCO-1.0-Signed-off-by: Ed Chejlava <ed.chejlava@lge.com>

Change-Id: Iccdba642c8b2a58a110143fd3b3fbefc827ad960
Reviewed-on: https://g2g.palm.com/2907
Reviewed-by: Build Verification
Reviewed-by: Ed Chejlava <ed.chejlava@lge.com>
Tested-by: Ed Chejlava <ed.chejlava@lge.com>
Reviewed-by: Roger Stringer <roger.stringer@lge.com>